extracting to specific class
